apti_backend/requirements.txt
aiohttp==3.7.3
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.3.1
asttokens==2.0.5
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.10.0
bio==0.8.5
biopython==1.79
biothings-client==0.2.6
bleach==4.1.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.11.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.6
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.2
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.8.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
dnspython==2.2.0
drf-yasg==1.20.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.0.12
firebase-admin==5.2.0
gcloud==0.18.3
geoip2==4.1.0
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth==2.1.0
google-auth-httplib2==0.1.0
google-cloud-core==2.0.0
google-cloud-firestore==2.3.3
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.40.0
gunicorn==20.0.4
h11==0.13.0
httplib2==0.19.1
idna==2.10
inflection==0.5.1
ipykernel==6.12.1
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.0.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.1
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
jwcrypto==1.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
maxminddb==2.0.3
mistune==0.8.4
MouseInfo==0.1.3
msgpack==1.0.2
multidict==5.0.2
mygene==3.2.2
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.21.2
oauth2client==4.1.3
openapi-codec==1.3.2
openpyxl==3.0.9
outcome==1.1.0
packaging==21.0
pandas==1.4.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
plac==1.3.3
prometheus-client==0.13.1
prompt-toolkit==3.0.29
proto-plus==1.19.0
protobuf==3.18.0
psutil==5.9.0
psycopg2==2.8.6
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.21
pycryptodome==3.14.1
PyGetWindow==0.0.9
Pygments==2.11.2
pymongo==4.0.1
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyperclip==1.8.2
PyRect==0.1.4
pyrsistent==0.18.1
PyScreeze==0.1.28
python-dateutil==2.8.2
python-jwt==3.3.2
pytube==10.9.0
pytube3==9.6.4
pytweening==1.0.4
pytz==2020.4
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
qrcode==7.3.1
qtconsole==5.3.0
QtPy==2.0.1
requests==2.25.0
requests-toolbelt==0.9.1
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
selenium==4.1.0
Send2Trash==1.8.0
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.1
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
virtualenv==20.4.7
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==5.2.0
widgetsnbextension==3.6.0
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.0.0
yarl==1.6.3
